Revert "Add import methods."
This reverts commit da70a9acea45af17b93cd30602b2f2b5b4ccb66a.
