+++ data-manager/api/src/main/java/org/orbisgis/orbisdata/datamanager/api/dataset/ITable.java
@@ -216,7 +216,7 @@ public interface ITable<T, U> extends IMatrix<T, U> {
      * @param filePath Path of the file to be saved.
      * @return True is the file has been saved, false otherwise.
      */
-    default boolean save(@Nullable String filePath) {
+    default boolean save(@NotNull String filePath) {
         return save(filePath, null);
     }
