+++ data-manager/api/src/test/java/org/orbisgis/orbisdata/datamanager/api/dsl/IOptionBuilderTest.java
@@ -37,7 +37,6 @@
 package org.orbisgis.orbisdata.datamanager.api.dsl;
 
 import org.junit.jupiter.api.Test;
-import org.orbisgis.orbisdata.datamanager.api.dsl.sql.IOptionBuilder;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
