+++ data-manager/jdbc/src/main/java/org/orbisgis/orbisdata/datamanager/jdbc/dsl/FromBuilder.java
@@ -34,19 +34,20 @@
  * or contact directly:
  * info_at_ orbisgis.org
  */
-package org.orbisgis.orbisdata.datamanager.jdbc.dsl.sql;
+package org.orbisgis.orbisdata.datamanager.jdbc.dsl;
 
 import org.orbisgis.orbisdata.datamanager.api.datasource.IJdbcDataSource;
-import org.orbisgis.orbisdata.datamanager.api.dsl.sql.IConditionOrOptionBuilder;
+import org.orbisgis.orbisdata.datamanager.api.dsl.IFromBuilder;
+import org.orbisgis.orbisdata.datamanager.api.dsl.IWhereBuilderOrOptionBuilder;
 import org.orbisgis.orbisdata.datamanager.jdbc.JdbcDataSource;
 
 /**
- * Implementation of {@link IConditionOrOptionBuilder}.
+ * Implementation of {@link IFromBuilder}.
  *
  * @author Erwan Bocher (CNRS)
  * @author Sylvain PALOMINOS (UBS 2019)
  */
-public class ConditionOrOptionBuilder extends OptionBuilder implements IConditionOrOptionBuilder {
+public class FromBuilder implements IFromBuilder {
 
     private final StringBuilder query;
     private final JdbcDataSource dataSource;
@@ -57,24 +58,15 @@ public class ConditionOrOptionBuilder extends OptionBuilder implements IConditio
      * @param request    String request coming from the ISelectBuilder.
      * @param dataSource {@link IJdbcDataSource} where the request will be executed.
      */
-    public ConditionOrOptionBuilder(String request, JdbcDataSource dataSource) {
-        super(request, dataSource);
+    public FromBuilder(String request, JdbcDataSource dataSource) {
         query = new StringBuilder();
         query.append(request).append(" ");
         this.dataSource = dataSource;
     }
 
     @Override
-    public IConditionOrOptionBuilder and(String condition) {
-        query.append("AND ");
-        query.append(condition);
-        return new ConditionOrOptionBuilder(query.toString(), dataSource);
-    }
-
-    @Override
-    public IConditionOrOptionBuilder or(String condition) {
-        query.append("OR ");
-        query.append(condition);
-        return new ConditionOrOptionBuilder(query.toString(), dataSource);
+    public IWhereBuilderOrOptionBuilder from(String... tables) {
+        query.append("FROM ").append(String.join(",", tables));
+        return new WhereBuilder(query.toString(), dataSource);
     }
 }
